requirements.txt
facenet_pytorch==2.6.0
fastapi==0.121.3
matplotlib==3.10.7
mediapipe==0.10.21
numpy==2.3.5
onnxruntime_gpu==1.19.2
opencv_contrib_python==4.11.0.86
Pillow==12.0.0
pydantic==2.12.4
qdrant_client==1.16.0
Requests==2.32.5
skimage==0.0
torch==2.2.2
tqdm==4.67.1
